fix upload, delete image; reup mail for resetpassword

+++ 04_Implement/FoodMap_Webservices/public_html/api/deletePicture.php
@@ -3,11 +3,18 @@ $respone = array();
 
 if (isset($_POST["url"]))
 {
-	$url = $_POST["url"];
-	unlink($url);
+	if (file_exists("../images/".$id)) {
+        $url = $_POST["url"];
+		unlink($url);
 
-	$respone["status"] = 200;
-	$respone["message"] = "Success";
+		$respone["status"] = 200;
+		$respone["message"] = "Success";
+    }
+    else
+    {
+    	$respone["status"] = 404;
+		$respone["message"] = "Not Found";
+    }
 }
 else
 {
@@ -16,5 +23,4 @@ else
 }
 
 echo json_encode($respone);
-
 ?>
